mob next [ci-skip] [ci skip] [skip ci]
lastFile:ui/src/pages/Projects/ProjectPage/NewChatSession/NewChatSession.tsx
